Rendering simplifications.
The only thing RootSurface did was track dirty cells for slightly optimised drawing, but only really introduced another layer of complexity.
Moving the dirty cell tracking into Surface lets every surface benefit from it.

The other change is letting Surface render itself to a given RenderTarget2D for when you want to put a Surface on a specific rendertarget for
effect passes or such.

Also reformatting because why not.